+++ tests/test_app/resources/store.py
@@ -1,12 +1,10 @@
-from typing import List
-
 import sqlalchemy as sa
 from marshmallow import fields
 
 import tests.test_app.models as m
 from awokado import custom_fields
 from awokado.consts import CREATE, READ, BULK_CREATE
-from awokado.utils import ReadContext, OuterJoin
+from awokado.utils import ReadContext
 from tests.test_app.resources.base import Resource
 
 
@@ -16,13 +14,13 @@ class StoreResource(Resource):
         name = "store"
         methods = (CREATE, BULK_CREATE, READ)
         auth = None
+        select_from = sa.outerjoin(
+            m.Store, m.Book, m.Store.id == m.Book.store_id
+        )
 
     id = fields.Int(model_field=m.Store.id)
     book_ids = custom_fields.ToMany(
-        fields.Int(),
-        resource="book",
-        model_field=m.Book.id,
-        join=OuterJoin(m.Store, m.Book, m.Store.id == m.Book.store_id),
+        fields.Int(), resource="book", model_field=m.Book.id
     )
     name = fields.String(model_field=m.Store.name, required=True)